ai-service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
python-dotenv==1.0.1
openai==1.35.0
django-cors-headers==4.4.0
gunicorn==22.0.0
